updates, remove APPRISS bot and should work!

+++ run.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 # Starts the IEMBot Process, run from ldm's crontab
 
-kill -9 `cat iembot.pid `
-sleep 5
+if [ -e iembot.pid ]; then
+	kill -9 `cat iembot.pid `
+	sleep 5
+fi
 twistd --logfile=logs/iembot.log --pidfile=iembot.pid -y iembot.tac
